+++ src/transformer/tags/helpers.ts
@@ -3,7 +3,7 @@ import { PageElement } from "@custom-site/page";
 import * as path from "path";
 
 /**
- * @param uri aタグのhref
+ * @param uri aタグの`href`, imgタグの`src`
  * @param page page.uriはoption.serverBasePathをすでに加算した状態で存在する
  * @param option 相対パスの算出では利用しない
  */
@@ -16,16 +16,16 @@ export const rewriteUrl = (uri: string, page: PageElement, option: CommonOption)
     calcUri = calcUri.slice(0, uri.length - "index".length);
   }
   if (uri.startsWith("/")) {
-    return path.join(option.serverBasePath, calcUri);
+    return path.join(option.basePath, calcUri);
   }
   if (uri.startsWith("../")) {
     const s = path.join(path.dirname(page.uri), calcUri);
     return s === "/" ? s : s.replace(/\/$/, "");
   }
   // start current directory
-  const t = path.join("/", path.dirname(page.uri), calcUri).replace(/\/$/, "");
-  if (option.serverBasePath !== "" && !t.startsWith(option.serverBasePath)) {
-    return path.join(option.serverBasePath, t);
+  const newUri = path.join("/", path.dirname(page.uri), calcUri).replace(/\/$/, "");
+  if (option.basePath !== "" && !newUri.startsWith(option.basePath)) {
+    return path.join(option.basePath, newUri);
   }
-  return t;
+  return newUri;
 };