+++ typings/@custom-site/index.d.ts
@@ -11,7 +11,8 @@ declare module "@custom-site/cli" {
     source: string;
     global: HtmlMetaProperties;
     destination?: string;
-    serverBasePath: string;
+    basePath: string;
+    port: number;
     blacklist: {
       extensions: string[];
     };
@@ -72,6 +73,12 @@ declare module "@custom-site/page" {
     globalLinks?: Array<string | LinkHTMLAttributes>;
   }
 
+  export interface ThirdParty {
+    googleAnalytics?: {
+      ua?: string;
+    };
+  }
+
   export interface HtmlMetaProperties extends OGP, TwitterMeta, ExternalJavaScript, ExternalCSS, ExternalLink {
     lang?: string;
     description?: string;
@@ -88,6 +95,7 @@ declare module "@custom-site/page" {
       "user-scalable"?: "no";
       width?: number | "device-width";
     };
+    thirdParty?: ThirdParty;
   }
 
   export type makeTemplateFunc = (props: PageProps) => (content?: React.ReactNode) => React.ReactElement<any>;